Reconstruct_Reader Utils: rotate crosshair-like traces by 45 degrees about their origin.
This commit is used to verify a boundary case bug in TrakEM2 alignment.

It appears that TrakEM2 will only apply transforms to Displayables with nonzero bounding box areas.
In other words, profiles (open contours in Reconstruct) that are colinear and either vertical or horizontal will remain in place across a least squares alignment.

The crosshair traces in Oblique aka r34js-gs are such traces. When aligning this series as-is, the crosshairs remain in place (ie, they no longer match the underlying EM).
This commit will rotate those traces by 45 degrees, after which, they will be transformed correctly.